# market_go_api_gateway/api/docs/swagger.yaml
definitions:
...
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  sale_service.Sale:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      payment_type:
        type: string
      price:
        type: number
      shop_assistant_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  sale_service.SaleCreateReq:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      payment_type:
        type: string
      price:
        type: number
      shop_assistant_id:
        type: string
      status:
        type: string
    type: object
  sale_service.SaleCreateResp:
    properties:
      msg:
        type: string
    type: object
  sale_service.SaleDeleteResp:
    properties:
      msg:
        type: string
    type: object
...
  sale_service.SaleUpdateReq:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      id:
        type: string
      payment_type:
        type: string
      price:
        type: number
      shop_assistant_id:
        type: string
      status:
        type: string
    type: object
  sale_service.SaleUpdateResp:
    properties:
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/sales:
    get:
      consumes:
      - application/json
      description: get sales
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      - description: search by branch_id
        in: query
        name: branch_id
        type: string
      - description: search by client_name
        in: query
        name: client_name
        type: string
      - description: search by payment_type
        in: query
        name: payment_type
        type: string
      - description: search by shop_assistant_id
        in: query
        name: shop_assistant_id
        type: string
      - description: search by cashier_id
        in: query
        name: cashier_id
        type: string
      - description: search by status
        in: query
        name: status
        type: string
      - description: search by created_at_from
        in: query
        name: created_at_from
        type: string
      - description: search by created_at_to
        in: query
        name: created_at_to
        type: string
      - description: search by price_from
        in: query
        name: price_from
        type: string
      - description: search by price_to
        in: query
        name: price_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sale_service.Sale'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: List sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Create a new sale with the provided details
      parameters:
      - description: data of the sale
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/sale_service.SaleCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sale_service.SaleCreateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new sale
      tags:
      - Sales
  /v1/sales/{id}:
    delete:
      consumes:
      - application/json
      description: delete a sale by its unique identifier
      parameters:
      - description: Sale ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.SaleDeleteResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a sale
      tags:
      - sales
    get:
      consumes:
      - application/json
      description: Retrieve a sale by its unique identifier
      parameters:
      - description: Sale ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.Sale'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a sale by ID
      tags:
      - Sales
    put:
      consumes:
      - application/json
      description: Update an existing sale with the provided details
      parameters:
      - description: Sale ID to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated data for the sale
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/sale_service.SaleUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale_service.SaleUpdateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing sale
      tags:
      - Sales
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"